Adding documents.company_id relation
@ntruchsess can you please have a look at the add on. With this we get the company_id into the documents table as NotNULL field.
We have a small duplication here, since in some cases, where the document is related to an consent (which will be the case in like 80% of the documents currently) the information of the user_id and the document_id will get stored in consent as well as documents. Maybe not ideal, but taking it out from consents makes no sense either. For me the best solution would be the one I just committed here. can you please run a quick test if everything is fine?